⚡️(api) improve performance of HTTP basic auth
HTTP Basic auth implementation in Ralph is using the secure and standard
bcrypt algorithm to hash/salt passwords before storing them.
But this implementation comes with a performance cost, as each request has an
overhead of almost 200ms.
By LRU caching the HTTP auth credentials, we should greatly decrease the
overhead of the vast majority of the requests.